refactor(backend): organize workflow modules
