+++ crates/backend/src/workflow/mod.rs
@@ -1,0 +1,6 @@
+//! Backend composition for workflow definitions and workflow runs.
+
+mod definition;
+pub(crate) mod run;
+
+pub(crate) use definition::WorkflowApi;
