+++ crates/backend/src/workflow/run/mod.rs
@@ -1,0 +1,14 @@
+//! Backend composition and runtime adapters for workflow runs.
+
+mod api;
+mod engine;
+mod executor;
+pub(crate) mod interactive;
+mod prerequisites;
+mod prompt;
+
+pub(crate) use api::WorkflowRunApi;
+pub(crate) use engine::{
+    ConcreteWorkflowRunControl, ConcreteWorkflowRunEngine, build_workflow_run_engine,
+    reconcile_running_workflow_runs,
+};
